feat!: widen price and tick size to u128 (3/3) (#122)
> **⚠️ Breaking change (Candid interface).** The `price`, `tick_size`,
and `lot_size` fields change from `nat64` to `nat`. Callers of
`add_limit_order` / `add_trading_pair` and consumers of
`TradingPairInfo`, `OrderRecord`, depth/ticker, and the event log must
update their candid bindings. Magnitudes that fit `nat64` encode
identically on the wire, so only the declared types change.

Final slice of the price-representation work, rebased on `main` (#119
and #121 have merged).

Widens the order-book `Price` key and `TickSize` from `u64` to
**`u128`** (Candid `nat`), so pairs whose per-whole-base price exceeds
`u64` are representable. An empirical check of live Binance spot pairs
found 22 that overflow `u64` under this denomination — all quoted in
18-decimal tokens (e.g. **BTC/FDUSD ≈ 6×10²²**).

## Performance

The wider key has no measurable cost in the matching engine (already
operating on `u256` quantities), confirming the earlier throwaway
measurement (experiment #118). Against `main`, the fill-heavy benchmarks
stay within ~1% (the 1000-order runs move +0.1%…+0.4%; the single large
fill +1.2%). The cost of the wider encoding is confined to small
absolute counts: cheap order-book read queries (depth/ticker, +~16–23%,
e.g. +134K on `depth_default`) and stable-structure upgrade
serialization (`upgrade_full_depth` +7.6%).

## PR stack

1. **#119** (merged) — establish and enforce the `tick_size × lot_size`
settlement-exactness invariant at pair creation.
2. **#121** (merged) — settle fills as `price × quantity /
10^base_decimals`. Stacked on #119.
3. **#122 (this PR)** — widen the price key to `u128` (Candid `nat`) for
high-value pairs. Stacked on #121.

---------

+++ canister/src/dashboard/tests.rs
@@ -213,7 +213,7 @@ fn place(
     state: &mut State<VectorMemory, VectorMemory>,
     user: Principal,
     side: Side,
-    price: u64,
+    price: u128,
     quantity: u64,
 ) -> OrderId {
     let pair: TradingPair = icp_ckbtc_trading_pair();
